pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cot-reasoner"
version = "0.1.0"
description = "A Chain of Thought Reasoner with multi-provider LLM support"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["llm", "chain-of-thought", "reasoning", "ai", "openai", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "openai",
    "anthropic",
    "fastapi",
    "uvicorn",
    "typer",
    "rich",
    "pydantic>=2.0",  # v2 required (breaking changes from v1)
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
]

[project.scripts]
cot-reasoner = "cot_reasoner.cli:app"
cot-reasoner-api = "cot_reasoner.api:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cot_reasoner"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
